完成登录校验，课程、教学、网关、用户模块全部通过测试

+++ therogram-api/src/main/java/com/wrathur/api/config/DefaultFeignConfig.java
@@ -17,7 +17,7 @@ public class DefaultFeignConfig {
         return new RequestInterceptor() {
             @Override
             public void apply(RequestTemplate template) {
-                Long userId = UserContext.getUser();
+                Integer userId = UserContext.getUser();
                 if (userId != null) {
                     template.header("user-info", userId.toString());
                 }
